RewardModel to own file
